Splitup MMR's `shared` functions into "basic" and "advanced"
The purpose of this is that only the functions in `shared_basic` needs
to be compiled to TASM, the rest do not.
