+++ twenty-first/src/util_types/mmr.rs
@@ -2,4 +2,5 @@ pub mod archival_mmr;
 pub mod mmr_accumulator;
 pub mod mmr_membership_proof;
 pub mod mmr_trait;
-pub mod shared;
+pub mod shared_advanced;
+pub mod shared_basic;
